requirements.txt
python-telegram-bot
fastapi
uvicorn
ollama
schedule
pytesseract
opencv-python
langchain
langchain-ollama
schedule

